Clean up members only used on UI thread
Follow up on the previous change (moving
```IInteractiveWindowOperations``` implementations to the UI thread).
